keep track of unrecognized postfix log entries
This can be useful if, for example, there are log entries that we're
unfamiliar with when writing the mtail program, but that come up
afterwards and turn out to be useful.

A good example might be the `postscreen` service or, in the case of
the Postfix exporter which I was using before, the "bounce" service:
the exporter wasn't tracking bounces, but thanks to that metric, that
could be deduced.