properly enable mtail on install
It seems like the mtail exporter was actually not enabled on
rdsys-test-01 and other hosts! I think that's an oversight of this
module, and flew totally off the radar until we enabled monitoring in
Prometheus because *eventually* those services got started by Puppet.

Now they should start on boot too.