stop installing pcregrep in mtail entirely
This is not necessary for mtail operation, and has been removed from
trixie. This overrides another approach (#2) where we instead install
it in older releases but trixie, but I felt that was just creating
technical debt, as we'd have to remove that code eventually.

Just bite the bullet and let downstreams deal with pcregrep as they
see fit.